requirements.txt
flask
pandas
spacy
nltk
rank-bm25
sutime
timex
aniso8601
bidaf-keras
tensorflow==1.13.1
absl-py==0.7.1
annoy==1.15.2
astor==0.7.1
fasteners==0.14.1
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lz4==2.1.6
Markdown==3.1
mock==2.0.0
monotonic==1.5
nltk==3.4.5
numpy==1.16.3
pbr==5.1.3
protobuf==3.7.1
pymagnitude==0.1.120
PyYAML==5.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==0.4.1
tqdm==4.31.1
Werkzeug==0.15.3
xxhash==1.3.0